🌷🍈 Checkpoint ./styles/more-collections.styl:433767/2052 ./src/components/collection/more-collections.js:433767/2642 ./src/presenters/collection-item.js:433767/4 ./src/presenters/more-collections.js:433767/3290
