+++ styles/more-collections.styl
@@ -17,22 +17,6 @@
   
     .collection-name-description
       margin-left: 0
-      
-// .more-collections-item
-//   border-radius: 5px
-//   display: flex
-//   flex-direction: column
-//   button
-//     margin: 12px
-//     width: fit-content
-//   span
-//     margin: 0 12px
-//     flex-grow: 1
-
-// .projects-count
-//   background-color: secondary-background
-//   padding: 12px
-//   border-bottom-right-radius: 5px
-//   border-bottom-left-radius: 5px
-//   align-self: flex-end
-//   width: 100%
+  
+  .projects-count
+    padding: 12px
